Added unit tests, refactored client on_message, created message_handler.py
* Added unit tests for 100% coverage in all files, except main
* Refactored on_message from main to MQTTClientWrapper, it only focuses
on handling the message and publishing it
* MQTTClientWrapper is passed message_handler on creation,
message_handler will handle validation, call necessary business logic,
and then return the data in an appropriate format
* Added validation to make sure int is not bool when checking if its non
negative

+++ rules_engine.py
@@ -1,11 +1,11 @@
 from constants import BASE_AMOUNTS, INELIGIBLE_AMOUNT, PER_CHILD_AMOUNTS
 
 
-def calculate_base_amount(program, family_composition, ):
-    return BASE_AMOUNTS[program][family_composition]
+def calculate_base_amount(program, family_composition):
+    return float(BASE_AMOUNTS[program][family_composition])
 
-def calculate_children_amount(program, number_of_children, ):  
-    return PER_CHILD_AMOUNTS[program] * number_of_children
+def calculate_children_amount(program, number_of_children):  
+    return float(PER_CHILD_AMOUNTS[program] * number_of_children)
 
 def calculate_supplement(program, number_of_children, family_composition, eligible):
     if not eligible:
